Add default meta tags to error page
